refactor(tests): #58 Tier 1.2 — split EODHD scaffold tests by feature
tests/test_eodhd_api_scaffold.py (1,060 LOC, 45 tests) →
tests/calendar/eodhd/{parsers,projector,client,fetcher,service_op}.py.

  test_parsers.py     18 tests — subtype registry + per-row parsers
                                 (earnings, trend, IPO, split, dividend,
                                 dividend_detail) — 327 LOC
  test_projector.py    4 tests — store_corp_raw / project_corp_events /
                                 v_calendar_item projection — 108 LOC
  test_client.py       4 tests — EODHDAPIClient auth + 429 retry +
                                 fmt-json injection — 91 LOC
  test_fetcher.py     14 tests — CorpCalendarFetcher dry-run / execution
                                 / dividend pagination + dividend_detail
                                 fetcher — 584 LOC
  test_service_op.py   5 tests — calendar_corp_fetch +
                                 calendar_corp_fetch_dividend_details — 94 LOC

Each split file imports only the symbols its tests reference and inlines
the row-builder helpers (`_earnings_row`, `_trend_row`, ...) plus the
`store` / `connection` fixtures it actually uses. Test bodies are
byte-equivalent to the original.

Verification: 45 tests pass on branch (same as master baseline).
